Place method to get albums in its component
